+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Specify PATCH, PUT, or MERGE method for EntityUpdateRequest - Barton Ip
 - Add a Service wide configuration (e.g. http.update\_method) - Jakub Filak
 - <, <=, >, >= operators on GetEntitySetFilter - Barton Ip
+- Django style filtering - Barton Ip
 
 ### Fixed
 - URL encode $filter contents - Barton Ip
-- Headers attribute on ODataHttpRequest - Barton Ip
+- JSON errors caused by invalid content length of Batch responses - Barton Ip
 
 ## [1.5.0]